backend/requirements.txt
Flask
flask-cors
pymongo
scikit-learn
numpy
pdfminer.six
python-docx
